Move umap run files to src folder such that toml file can run
