+++ pyproject.toml
@@ -30,8 +30,7 @@ dependencies = [
 # the documentation) but not necessarily required for _using_ it.
 dev = [
     "flake8",
-    "pre-commit",
-    "pytest-mock",
+    "pre-commit"
 ]
 
 [project.urls]
